recreated workspace to resume from last saved version. edited migration files to implement changes done in other migration files, hence reducing number of migration files. performed migration up to 20220811080929

+++ migrations/20220810092551-tea-type.js
@@ -16,7 +16,7 @@ exports.setup = function(options, seedLink) {
 
 exports.up = function(db) {
   return db.createTable('tea_types',{
-    id:{type:'tinyint',unsigned:true,primaryKey:true,autoIncrement:true},
+    id:{type:'smallint',unsigned:true,primaryKey:true,autoIncrement:true},
     name:{type:'string',length:20}
   });
 };
